+++ fractal/mandelbrot.h
@@ -13,7 +13,7 @@ struct Mandelbrot {
     Mandelbrot(int width, int height);
 
     void resize(int width, int height);
-    uint32_t const* generate(float x, float y, float zoom);  ///< pointer may be invalidated after any interaction with this object
+    uint32_t const* generate(double x, double y, double zoom);  ///< pointer may be invalidated after any interaction with this object
 
 private:
     std::vector<uint32_t> pixels_;
@@ -24,7 +24,7 @@ private:
             std::make_unique<ThreadPool>(std::thread::hardware_concurrency());
     std::queue<std::future<void>> futures_;
 
-    uint32_t color(float x, float y, int iterations, int max_iterations);
+    uint32_t color(double x, double y, int iterations, int max_iterations);
 };
 
 }  // namespace fractal
